Don't recurse into embedded variants
Embedded variants are not part of the type, but are usually opaque data
that the site dealing with the surrounding variant knows nothing about.

Treat them as such by filling a `CVariant **` when encountering a 'v' in
c_variant_read() and acception a `CVariant *` in c_variant_write().

This makes it easier to pass around embedded variants.